Refac and add collapsable forecast widget
