+++ services/frontend/src/App.js
@@ -1,41 +1,12 @@
-import WeatherWidgetContainer from './components/WeatherWidgetContainer';
-import Header from './components/Header';
+import WeatherForecast from './components/WeatherForecast';
 
 export const App = () => {
-    const locationData = {
-        city: "San Francisco",
-        state: "CA",
-    }
-
-    const weatherData = [
-        {
-            temp: 51,
-            cond: 'cloudy',
-        },
-        {
-            temp: 48,
-            cond: 'rain',
-        },
-        {
-            temp: 58,
-            cond: 'windy',
-        },
-        {
-            temp: 62,
-            cond: 'partial-cloudy',
-        },
-        {
-            temp: 71,
-            cond: 'sunny',
-        },
-    ];
 
     return (
-        <div className="flex">
-            <div className="md:w-1/2 mx-auto my-5 rounded-xl shadow-md text-gray-700 text-center text-xl p-4 bg-gradient-to-b from-blue-300 to-blue-50">
-                <Header city={locationData.city} state={locationData.state} />
-                <WeatherWidgetContainer weatherObjArray={weatherData} />
-            </div>
+        <div>
+            <WeatherForecast locationData={{city: 'San Francisco', state: 'CA'}} defaultExpanded={true}/>
+            <WeatherForecast locationData={{city: 'Chicago', state: 'IL'}}/>
+            <WeatherForecast locationData={{city: 'Oakland', state: 'CA'}}/>
         </div>
     )
 }
